+++ tsconfig.json
@@ -25,7 +25,9 @@
   ],
   "include": [
     "next-env.d.ts",
+    "**/*.js",
+    "**/*.jsx",
     "**/*.ts",
     "**/*.tsx"
-  ]
+, "components/layout/SectionProvider.jsx"  ]
 }
